no allocs, more stack
