+++ src/Paprika/Merkle/CommitExtensions.cs
@@ -1,0 +1,35 @@
+using Paprika.Chain;
+using Paprika.Data;
+
+namespace Paprika.Merkle;
+
+/// <summary>
+/// Extensions to the commit used to set values of the 
+/// </summary>
+public static class CommitExtensions
+{
+    public static void SetLeaf(this ICommit commit, in Key key, in NibblePath leafPath)
+    {
+        var leaf = new Node.Leaf(leafPath);
+        commit.Set(key, leaf.WriteTo(stackalloc byte[leaf.MaxByteLength]));
+    }
+
+    public static void SetBranch(this ICommit commit, in Key key, NibbleSet.Readonly children,
+        NibbleSet.Readonly dirtyChildren)
+    {
+        var branch = new Node.Branch(children, dirtyChildren);
+        commit.Set(key, branch.WriteTo(stackalloc byte[branch.MaxByteLength]));
+    }
+
+    public static void SetBranchAllDirty(this ICommit commit, in Key key, NibbleSet.Readonly children)
+    {
+        var branch = new Node.Branch(children, children);
+        commit.Set(key, branch.WriteTo(stackalloc byte[branch.MaxByteLength]));
+    }
+
+    public static void SetExtension(this ICommit commit, in Key key, in NibblePath path)
+    {
+        var extension = new Node.Extension(path);
+        commit.Set(key, extension.WriteTo(stackalloc byte[extension.MaxByteLength]));
+    }
+}
